pyproject.toml
[tool.poetry]
name = "update_state_list"
version = "0.00"
description = "Utility for updating state list"
authors = ["gbabineau <guy.babineau@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.2.0"
requests = "^2.31.0"
ebird-api = "^3.0.6"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
coverage = "^7.3.1"
pytest = "^9.0.1"
pytest-cov = "^4.1.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source=["."]
branch = true
command_line = "-m pytest --junitxml=junit/test-results.xml"
relative_files = true

[tool.coverage.report]
show_missing = true

[tool.coverage.xml]
output = "coverage.xml"

[tool.poetry.group.lint.dependencies]
ruff = "^0.11.6"
mypy = "^1.11.2"

[tool.ruff]
line-length = 80

